requirements.txt
flask
python-dotenv
pandas
google-generativeai